Split SvgLinearGradientServer
